pyproject.toml
[tool.poetry]
name = "lsu-pilot"
version = "0.1.0"
description = ""
authors = ["Pedro Javier Martos Velasco <pmartos@theworkshop.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.52.1"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.6"
pandas = "^2.2.3"
tiktoken = "^0.8.0"
langchain = "^0.3.4"
numpy = "^1.21"
scipy = "^1.14.1"
cairosvg = "^2.7.1"
requests = "^2.32.3"
whisper-cpp-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
